requirements.txt
Flask==3.1.3
Flask-SQLAlchemy==3.1.1
Flask-Mail==0.9.1
Flask-Cors==6.0.2
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.7.1
Werkzeug==3.1.6
itsdangerous==2.2.0
Authlib==1.6.6
reportlab==4.0.7
openpyxl==3.1.2
gunicorn==22.0.0
psycopg2-binary==2.9.10
python-dotenv==1.0.0
requests==2.32.4
pytz==2024.1
